fix: refactor ante package, add deprecated msgs
