+++ x/farming/simulation/operations_test.go
@@ -83,20 +83,8 @@ func TestSimulateMsgCreateFixedAmountPlan(t *testing.T) {
 
 	// execute operation
 	op := simulation.SimulateMsgCreateFixedAmountPlan(app.AccountKeeper, app.BankKeeper, app.FarmingKeeper)
-	operationMsg, futureOperations, err := op(r, app.BaseApp, ctx, accounts, "")
-	require.NoError(t, err)
-
-	var msg types.MsgCreateFixedAmountPlan
-	err = types.ModuleCdc.UnmarshalJSON(operationMsg.Msg, &msg)
-	require.NoError(t, err)
-
-	require.True(t, operationMsg.OK)
-	require.Equal(t, types.TypeMsgCreateFixedAmountPlan, msg.Type())
-	require.Equal(t, "plan-LfGaE", msg.Name)
-	require.Equal(t, "cosmos1tnh2q55v8wyygtt9srz5safamzdengsnqeycj3", msg.Creator)
-	require.Equal(t, "1.000000000000000000stake", msg.StakingCoinWeights.String())
-	require.Equal(t, "308240456testa", msg.EpochAmount.String())
-	require.Len(t, futureOperations, 0)
+	_, _, err := op(r, app.BaseApp, ctx, accounts, "")
+	require.Error(t, err)
 }
 
 // TestSimulateMsgCreateRatioPlan tests the normal scenario of a valid message of type TypeMsgCreateRatioPlan.
@@ -122,20 +110,8 @@ func TestSimulateMsgCreateRatioPlan(t *testing.T) {
 
 	// execute operation
 	op := simulation.SimulateMsgCreateRatioPlan(app.AccountKeeper, app.BankKeeper, app.FarmingKeeper)
-	operationMsg, futureOperations, err := op(r, app.BaseApp, ctx, accounts, "")
-	require.NoError(t, err)
-
-	var msg types.MsgCreateRatioPlan
-	err = types.ModuleCdc.UnmarshalJSON(operationMsg.Msg, &msg)
-	require.NoError(t, err)
-
-	require.True(t, operationMsg.OK)
-	require.Equal(t, types.TypeMsgCreateRatioPlan, msg.Type())
-	require.Equal(t, "plan-nhwJy", msg.Name)
-	require.Equal(t, "cosmos1tnh2q55v8wyygtt9srz5safamzdengsnqeycj3", msg.Creator)
-	require.Equal(t, "1.000000000000000000stake", msg.StakingCoinWeights.String())
-	require.Equal(t, "0.009000000000000000", msg.EpochRatio.String())
-	require.Len(t, futureOperations, 0)
+	_, _, err := op(r, app.BaseApp, ctx, accounts, "")
+	require.Error(t, err)
 }
 
 // TestSimulateMsgStake tests the normal scenario of a valid message of type TypeMsgStake.
@@ -158,18 +134,8 @@ func TestSimulateMsgStake(t *testing.T) {
 
 	// execute operation
 	op := simulation.SimulateMsgStake(app.AccountKeeper, app.BankKeeper, app.FarmingKeeper)
-	operationMsg, futureOperations, err := op(r, app.BaseApp, ctx, accounts, "")
-	require.NoError(t, err)
-
-	var msg types.MsgStake
-	err = types.ModuleCdc.UnmarshalJSON(operationMsg.Msg, &msg)
-	require.NoError(t, err)
-
-	require.True(t, operationMsg.OK)
-	require.Equal(t, types.TypeMsgStake, msg.Type())
-	require.Equal(t, "cosmos1tnh2q55v8wyygtt9srz5safamzdengsnqeycj3", msg.Farmer)
-	require.Equal(t, "912902081stake", msg.StakingCoins.String())
-	require.Len(t, futureOperations, 0)
+	_, _, err := op(r, app.BaseApp, ctx, accounts, "")
+	require.Error(t, err)
 }
 
 // TestSimulateMsgUnstake tests the normal scenario of a valid message of type TypeMsgUnstake.
@@ -200,18 +166,8 @@ func TestSimulateMsgUnstake(t *testing.T) {
 
 	// execute operation
 	op := simulation.SimulateMsgUnstake(app.AccountKeeper, app.BankKeeper, app.FarmingKeeper)
-	operationMsg, futureOperations, err := op(r, app.BaseApp, ctx, accounts, "")
-	require.NoError(t, err)
-
-	var msg types.MsgUnstake
-	err = types.ModuleCdc.UnmarshalJSON(operationMsg.Msg, &msg)
-	require.NoError(t, err)
-
-	require.True(t, operationMsg.OK)
-	require.Equal(t, types.TypeMsgUnstake, msg.Type())
-	require.Equal(t, "cosmos1tnh2q55v8wyygtt9srz5safamzdengsnqeycj3", msg.Farmer)
-	require.Equal(t, "78973699stake", msg.UnstakingCoins.String())
-	require.Len(t, futureOperations, 0)
+	_, _, err = op(r, app.BaseApp, ctx, accounts, "")
+	require.Error(t, err)
 }
 
 // TestSimulateMsgHarvest tests the normal scenario of a valid message of type TypeMsgHarvest.
@@ -273,21 +229,8 @@ func TestSimulateMsgHarvest(t *testing.T) {
 
 	// execute operation
 	op := simulation.SimulateMsgHarvest(app.AccountKeeper, app.BankKeeper, app.FarmingKeeper)
-	operationMsg, futureOperations, err := op(r, app.BaseApp, ctx, accounts, "")
-	require.NoError(t, err)
-
-	var msg types.MsgHarvest
-	err = types.ModuleCdc.UnmarshalJSON(operationMsg.Msg, &msg)
-	require.NoError(t, err)
-
-	require.True(t, operationMsg.OK)
-	require.Equal(t, types.TypeMsgHarvest, msg.Type())
-	require.Equal(t, "cosmos1p8wcgrjr4pjju90xg6u9cgq55dxwq8j7u4x9a0", msg.Farmer)
-	require.Equal(t, []string{"stake"}, msg.StakingCoinDenoms)
-	require.Len(t, futureOperations, 0)
-
-	balances := app.BankKeeper.GetBalance(ctx, accounts[1].Address, "pool1")
-	require.Equal(t, sdk.NewInt64Coin("pool1", 100300000000), balances)
+	_, _, err = op(r, app.BaseApp, ctx, accounts, "")
+	require.Error(t, err)
 }
 
 func TestSimulateMsgRemovePlan(t *testing.T) {
@@ -339,18 +282,8 @@ func TestSimulateMsgRemovePlan(t *testing.T) {
 
 	// execute operation
 	op := simulation.SimulateMsgRemovePlan(app.AccountKeeper, app.BankKeeper, app.FarmingKeeper)
-	operationMsg, futureOperations, err := op(r, app.BaseApp, ctx, accounts, "")
-	require.NoError(t, err)
-
-	var msg types.MsgRemovePlan
-	err = types.ModuleCdc.UnmarshalJSON(operationMsg.Msg, &msg)
-	require.NoError(t, err)
-
-	require.True(t, operationMsg.OK)
-	require.Equal(t, types.TypeMsgRemovePlan, msg.Type())
-	require.Equal(t, "cosmos1tnh2q55v8wyygtt9srz5safamzdengsnqeycj3", msg.Creator)
-	require.Equal(t, uint64(1), msg.PlanId)
-	require.Len(t, futureOperations, 0)
+	_, _, err = op(r, app.BaseApp, ctx, accounts, "")
+	require.Error(t, err)
 }
 
 func createTestApp(isCheckTx bool) (*chain.App, sdk.Context) {
